GameScenes, scene transitions, audio transfer between states, music fadein, extra Rect ops, slight changes on screenfade and controller support implemented, DialogueBox in process of refinement, stopped when trying to use parallax on sprites

+++ include/MathTools.h
@@ -8,6 +8,6 @@ double WrapMinMax(double Val, double Min, double Max);//Val -> [Min,Max)
 
 double Clamp(double x, double Lower=0, double Higher=1); 
 double SmoothStep (double LowerBound, double HigherBound, double x);
-double SmootherStep(float LowerBound, double HigherBound, double x);
+double SmootherStep(double LowerBound, double HigherBound, double x);
 
 #endif// MATH_EXTRATOOLS
